# FitApp/FitApp/Models/FoodItemsModel.cs
namespace FitApp.Models
{
    public class FoodItemsModel
    {
        public int FoodItemID { get; set; }
        public string FoodItemName { get; set; }
        public string FoodItemDescription { get; set; }
        public string FoodItemCalories { get; set; }
    }
}