# FitApp/FitApp/Models/MealFoodItemsModel.cs
namespace FitApp.Models
{
    public class MealFoodItemsModel
    {
        public int MyProperty { get; set; }
        public int FoodItemID { get; set; }
        public int MealID { get; set; }
        public string ServingSize { get; set; }
        public string ServingsPerMeal { get; set; }
    }
}