# FitApp/FitApp/Models/MealsModel.cs
namespace FitApp.Models
{
    public class MealsModel
    {
        public int MealID { get; set; }
        public string MealName { get; set; }
        public string MealDescription { get; set; }
        public string MealCalories { get; set; }
        public int UserID { get; set; }
    }
}