# FitApp/FitApp/Models/UsersModel.cs
namespace FitApp.Models
{
    public class UsersModel
    {
        public int UserID { get; set; }
        public string UserName { get; set; }
        public string Password { get; set; }
        public string Email { get; set; }
        public string FirstName { get; set; }
        public string LastName { get; set; }
        public string Gender { get; set; }
        public int Age { get; set; }
        public decimal Weight { get; set; }
        public decimal Height { get; set; }
        public string ActivityLevel { get; set; }
    }
}