# FitApp/FitApp/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FitApp.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/FoodItems/{id}": {
      "get": {
        "tags": [
          "FoodItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FoodItems"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodItems"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodItems"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "FoodItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoodItems"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FoodItems"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FoodItems"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "FoodItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/MealFoodItems": {
      "get": {
        "tags": [
          "MealFoodItems"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealFoodItems"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealFoodItems"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealFoodItems"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MealFoodItems"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealFoodItems"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MealFoodItems"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MealFoodItems"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MealFoodItems"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealFoodItems"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealFoodItems"
                }
              }
            }
          }
        }
      }
    },
    "/api/MealFoodItems/{id}": {
      "get": {
        "tags": [
          "MealFoodItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MealFoodItems"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealFoodItems"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealFoodItems"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "MealFoodItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealFoodItems"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MealFoodItems"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MealFoodItems"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "MealFoodItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Meals/{id}": {
      "get": {
        "tags": [
          "Meals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Meals"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meals"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meals"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Meals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Meals"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Meals"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Meals"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Meals"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FoodItems": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "modificationDate": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "foodItemID": {
            "type": "integer",
            "format": "int32"
          },
          "foodItemName": {
            "type": "string",
            "nullable": true
          },
          "foodItemDescription": {
            "type": "string",
            "nullable": true
          },
          "foodItemCalories": {
            "type": "string",
            "nullable": true
          },
          "mealFoodItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MealFoodItems"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MealFoodItems": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "modificationDate": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "mealFoodItemId": {
            "type": "integer",
            "format": "int32"
          },
          "foodItemID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "foodItem": {
            "$ref": "#/components/schemas/FoodItems"
          },
          "mealID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "meal": {
            "$ref": "#/components/schemas/Meals"
          },
          "servingSize": {
            "type": "string",
            "nullable": true
          },
          "servingsPerMeal": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Meals": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "modificationDate": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "mealID": {
            "type": "integer",
            "format": "int32"
          },
          "mealName": {
            "type": "string",
            "nullable": true
          },
          "mealDescription": {
            "type": "string",
            "nullable": true
          },
          "mealCalories": {
            "type": "string",
            "nullable": true
          },
          "userID": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/Users"
          },
          "mealFoodItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MealFoodItems"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Users": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "creationDate": {
            "type": "string",
            "format": "date-time"
          },
          "modificationDate": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "userID": {
            "type": "integer",
            "format": "int32"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "weight": {
            "type": "number",
            "format": "double"
          },
          "height": {
            "type": "number",
            "format": "double"
          },
          "activityLevel": {
            "type": "string",
            "nullable": true
          },
          "workoutPlans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Users"
            },
            "nullable": true
          },
          "meals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Meals"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}